♻️  [Refactor] NetworkManager 타입 내 메서드 리팩터링
- 인코딩 시 모델 타입을 딕셔너리 형식으로 변환하는 asDictionary() 메서드 추가
- URLRequest 초기화와 함께 HTTPMethod를 세팅해주는 set(url: httpMethod:) 추가
- multipart/form-data 형식으로 변환하는 MultipartConvertible 프로토콜 초기구현
- delete, post, patch HTTPMethod 요청을 만드는 Requestable 프로토콜 초기구현
- HTTPMethod 타입을 파일로 분리
